Import of commit 339913cbf0268da82ab517cfdbf5e3c3ece99191 by Ry Biesemeyer <yaauie@users.noreply.github.com>
acks: send sequence-0 ack for 0-window batches (#479)

* acks: send sequence-0 ack for 0-window batches

Ensures that at least one ACK is sent in reply to each window, including
those with 0-length, by emitting an empty batch when we receive a 0-window
frame and sending a sequence-0 ACK in reply when we process an empty batch.

This enables upstream users of the protocol to (ab)use a 0-length window
frame and an _ensured_ in-protocol response to validate that we are alive.

Note: 0-sequence ACKs are already used as application-level keep-alives
between when the `ConnectionHandler` receives bytes and when `BeatsHandler`
finishes processing a batch, and therefore should already be handled
gracefully by any existing clients.